some more rework around the SampleFile
