+++ src/readme.ts
@@ -1,19 +1,14 @@
-import { SampleFile } from './file';
 import { Project } from './project';
+import { SampleFile } from './sample-file';
 
 /**
  * Represents a README.md file. Unlike most other files this will only be created if the file doesn't already exist.
  * You are expected to manage this file after creation.
+ *
+ * @param text - The initial contents of the README.md file. Defaults to '# replace this'
  */
 export class Readme extends SampleFile {
-  private readonly text: string;
-
   constructor(project: Project, text?: string) {
-    super(project);
-    this.text = text ?? '# replace this';
-  }
-
-  synthesize(outdir: string) {
-    this.writeOnceFileContents(outdir, 'README.md', this.text);
+    super(project, 'README.md', { contents: text ?? '# replace this' });
   }
 }
